pass in float position

+++ hs_detection/detect/Detection.cpp
@@ -8,7 +8,7 @@ namespace HSDetection
 {
     Detection::Detection(int numChannels, int chunkSize, int chunkLeftMargin,
                          int spikeLen, int peakLen, int threshold, int minAvgAmp, int maxAHPAmp,
-                         int *channelPositions, float neighborRadius, float innerRadius,
+                         float *channelPositions, float neighborRadius, float innerRadius,
                          int noiseDuration, int spikePeakDuration,
                          bool decayFiltering, float noiseAmpPercent,
                          bool localize, int numCoMCenters,
